Criação do .gitgnore e pastas
