+++ src/Empregado/Engenheiro.cpp
@@ -1,0 +1,11 @@
+#include <string>
+#include "Empregado.hpp"
+
+class Engenheiro : public Empregado {
+
+  public:
+    std::string nome;  
+	int projetos;
+	
+};
+
